Various fixes for the wy boogie command

+++ src/main/java/wy2boogie/util/Boogie.java
@@ -24,6 +24,7 @@ import java.util.Map;
 import java.util.concurrent.TimeUnit;
 
 import wy2boogie.core.BoogieFile;
+import wybs.lang.SyntacticException;
 import wyfs.util.ArrayUtils;
 
 /**
@@ -78,12 +79,13 @@ public class Boogie {
 	 * @return
 	 */
 	public Error[] check(int timeout, BoogieFile boogie) {
+		String filename = null;
 		try {
 			// Determine id for the temporary file. This will help to identified should
 			// there be a need during debugging.
 			String id = boogie.getEntry().id().toString();
 			// Create the temporary file.
-			String filename = createTemporaryFile(id, ".bpl", boogie.getBytes());
+			filename = createTemporaryFile(id, ".bpl", boogie.getBytes());
 			// ===================================================
 			// Construct command
 			// ===================================================
@@ -113,20 +115,24 @@ public class Boogie {
 				boolean success = child.waitFor(timeout, TimeUnit.MILLISECONDS);
 				byte[] stdout = readInputStream(input);
 				byte[] stderr = readInputStream(error);
+				System.out.println("STDOUT: " + new String(stdout));
+				System.out.println("STDERR: " + new String(stdout));
 				if(success && child.exitValue() == 0) {
 					return parseErrors(new String(stdout));
 				}
-				System.out.println("STDOUT: " + new String(stdout));
 			} finally {
 				// make sure child process is destroyed.
 				child.destroy();
+			}
+		} catch(IOException e) {
+			throw new RuntimeException(e.getMessage(), e);
+		} catch(InterruptedException e) {
+			throw new RuntimeException(e.getMessage(),e);
+		} finally {
+			if(filename != null) {
 				// delete the temporary file
 				new File(filename).delete();
 			}
-		} catch(IOException e) {
-
-		} catch(InterruptedException e) {
-
 		}
 		return null;
 	}
